Move test-only data generation to the test module
